- Spliting code in multiple classes for better understanding
